+++ .gitignore
@@ -15,3 +15,4 @@ docs/_build
 .Rhistory
 .ionide
 gs-testing-stuff
+.mypy.cache
