+++ otter/test_files/__init__.py
@@ -1,0 +1,2 @@
+from .abstract_test import TestCollection
+from .ok_test import OKTestFile
